+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [vNext]
+- Fixed attributes in build tasks
 
 ## [0.17.0] / 2019-02-23
 - Removed collection-based tasks in `FileSystemTasks`